Update README to warn against use of Negative Captcha with caching
I could be wrong about this, but I just spent a bit of time looking through the source code here and trying to debug an issue on my production servers, and the conclusion that I came to was that it was caused by caching. If I'm correct, I figure this could be worth including here to save others from running into the same issue. Big fan of the concept and gem.